LayoutTests/js/dom/modules/module-src-dynamic-in-xhtml.xhtml

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test dynamically added module with "src" attribute does not work in XHTML document.');

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
<script>
debug('Module is not executed yet.');
(function () {
    var element = document.createElement('script');
    element.type = 'module';
    element.src = './script-tests/module-src-dynamic.js';
    document.body.appendChild(element);
    setTimeout(function () {
        shouldBe(`window.exportedCocoa`, `undefined`);
        finishJSTest();
    }, 100);
}());
</script>
</body>
</html>
