Safari-611.1.10/LayoutTests/js/dom/modules/module-src-dynamic-in-xhtml-expected.txt

Test dynamically added module with "src" attribute does not work in XHTML document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS window.exportedCocoa is undefined
PASS successfullyParsed is true

TEST COMPLETE
